Add day_number method
This uses similar math to what is used to calculate a [Julian
Day](https://en.wikipedia.org/wiki/Julian_day). The math is best
described here:
http://www.cs.utsa.edu/~cs1063/projects/Spring2011/Project1/jdn-explanat
ion.html but let's walk through it.

Note: I'll explain it with month names from the Gregorian calendar, but
variables are named with their Golarion equivalents. Golarion's
calendar has an equivalent month order, month lengths, year lengths,
and has a leap year only and always every 8 years. A bit simpler than
Gregorian.

It works on the basis of a calendar that starts in March and ends in
February. This allows leap years to happen right at the end of a year
instead of sometime in the middle.

This is accomplished by shifting all the month's by 9 then "mod 12"-ing
them. This puts March at 0, April at 1 ... January at 10, February at
11. This is called a shifted_month.

Then we figured out the days in previous months. March is month 0 so if
we're in June, we'll be calculating all the days in March, April, and
May. I'll admit, this uses a magic number that I don't totally
understand, but I know it works. One thing that's important it all
division here is integer division, for example 5/3 is rounded down to 1.

Adding on the number of days in the current month, we have the number
of days since the start of the shifted year. We'll come back to this.

We then calculate if the non-shifted month is Jan or Feb represented by
1 for Jan or Feb and 0 for the other 10 months. We're dealing with
shifted months (and therefore years) remember? This subtracts a year if
we're in January or February because the shifted year hasn't started
yet.

Next, two more understandable calculations; getting all the days in
normal years and all the leap days up until now. Note: integer division
still in effect for leap days. 8 years / 8 is 1 and 11 years / 8 still
is 1.

We then have to un-offset our days because of the weird month shift. We
take a normal year and remove the days from Jan and Feb which is 306
days.

Then we add all the days in normal years, days in previous months, leap
days, and current day of the month, and find we're off by 306. Great.
But that's why we got the offset days in the previous step so we can
subtract them off.

And that's how you get the Golarion day number.