Failed attempts at parsing date from day_number
This was an attempt to derive an ArDate from a day number. For example
day 1 is the same as ArDate.new(year: 1, month: 1, day: 1) It was
loosely based around a similar algorithm used to convert the opposite
direction (year/month/day to number of days since year 0), but it was
very gross looking and didn't work well.

This is kept here as a historical artifact.

progress
'ar-date-parser' on '324f60b'.