Refactor ArDateParser.from_day_number
This uses the same logic as the previous commit, but I'll restate it
here for posterity:

This algorithm is used to calculate the year, month, and day from the
number of days since the first day. The first day is year 1, month 1,
day 1.

First we figure out the year. Since leap days are every 8 years, the
average length of a year is 365 and 1/8, or 365.125 days. Dividing
day_number by an average year gives us a fractional year. For example,
30 days / 365.125 is 0.082 years. Rounded up to the nearest integer
this is 1, meaning we are in year 1.

The average year comes in very handy for leap years. On the last day of
year 8 (the first leap year), the day number is 2921. 2921 divided by
365.125 is exactly 8. That means the next day will round up to 9
because it is, in fact, in year 9.

Second, we figure out the days into the current year. We figure out how
many days came before the current year. We do this by multiplying 365
by the number of years - 1, then we add in a leap day for every 8 years
(not counting the current year). Once we have the number of days before
the year started, we subtract that from day_number.

Month start days are the day in the year before each month starts.
Month 1 starts at 0 days, Month 2: 31 days, Month 3: 59 days, etc.
These are used the upcoming calculations.

If we're in a leap year, we use a different set of month start days
because everything after month 2 is shifted forward by the leap day. We
know if we're in a leap year by checking if the current year is
divisible by 8 evenly.

Using the number of days into the year, we can figure out the month and
day.

The month starts at 0, iterates through each month to see if we're past
the start day of that month. If we aren't, it adds 1 to the month until
we find the current month.

To find the day, we subtract the start day of that month from the
day_number, giving us the days left over in that month.

Ta-da!

'ar-date-parser' on '317b0b1'.