Fix bug in ArDateParser month finding
Previously, when trying to find the month, we looped through each
month's start day and compared it to the number of days into a year. If
the day was greater to the month start day we knew we were at least in
that month, if not later, and so we added one to the month. When the
month start day finally was more than the day of the year, it broke the
loop and stopped adding to the month. We were then left with a the last
month that the day was in.

The problem was when the month was the last month of the year. The loop
would get to the last month, and of course our day is in that month so
month would now be equal to 12. It would then try to find the next
month, return nil, and crash.

This change instead selects all months whose start days are less than
the day of the year. We're left with a list of months that we can count
and use for the current month number.

fixes #5