Refactor ArDate day_number method
This breaks up the method into a few smaller, more readable methods. It
also removes a bunch of magic numbers.

The way the day_number gets calculated now is by adding:

* Days in the previous years (discounting leap days)
* Leap days in the previous years
* The days of the year until the beginning of this month
* The current day of the month

For the month days, it uses a different set of days depending on if
we're currently in a leap year or not.

And that's how it works! So much simpler.