Add month addition to ArDate
The motivation for adding this is to be able create a link to the next
month in the calendar. The core of this logic is the divmod method
which returns both the result of the division and modulo in an array.
We add the current month number to the incoming number of months, then
divmod(12) and get the resulting new month, and the years to add to the
current year.

An additional adjustment is needed for months where the day would be
invalid. For example, starting in the third month at day 31 and trying
to advance one month to month two which has a max of 30 days. We do a
final check to make sure it's not more than that month's max day count.

on 'dd3dee4'.