Add argument to ArDate.beginning_of_week
Before, this would always been moonday, but now it can be set via this
single argument.

The code ended up a little kludgey. First, it gets the index of the
desired start day. Second, it figures out how many days until the next
start day. This is the part I don't like. There is a magic number
(days per week) that is used to make sure we don't end up with a
negative number.

Finally, we invert the days to get the distnace in days to the start of
the week. Subtracting that from the current day number gives us a new
day number which we can create a valid date out of.

I'd like to refactor some of that soon. Probably the ArDate constants as
I'm using them much differently than they are structured. That will
probably be soon in future commits.