Refactor ArDate
This is based around the refactoring of beginning_of_week and
end_of_week. I took the opportunity to restructure the constants and
turning them into arrays instead of hashes because that's how they were
mostly accessed.

For the refactoring of beginning_ and end_of_week, I was able to clean
up the logic to have it be more readable. end_of_week references the
last day of the week instead of the first, for example. Also, there are
far less magic numbers.

Here's how they work now:

First we get the index of the first or last day of the week for
`beginning_of_week` or `end_of_week` respectively. I'll call this the
week's anchor. Second, we get the difference in days between the current
day and the week's anchor. Third, we add or subtract this difference to
the date's `day_number` to create a new day number. Finally, this final
number is used to create a new date.