Add to_s to ArDate
This is essentially a way to convert a date into a string in the same
way the ArDateParser.from_date_string converts a string into a date.
This will be used in a query parameter to pick the date to display on
screen.

It's zero-padded because that's what Ruby's Date .to_s does.

fixes #16