Fix month math subtraction
When moving from month 1 backwards one month to month 12, the month
would be reported as 0 not 12. This is because months are 1-indexed so
(month 1) - (1 month) is 0.

The fix is the >> method is to always subtract 1 from the current month
to make it zero indexed. I'll call this month_index. In an array from
1..12, this month_index points to the correct month number.