Create basic event model
I decided to store events by the day number they occurred_on along with
a title. I went with day number because there's no ArDate type for
postgres. To bridge this gap, I used the ActiveRecord Attributes API to
convert between a the day number and an ArDate.

Another consideration I had was to switch ArDate over to just store its
day_number instead of year/month/day. I may still do that but I haven't
pulled the trigger. At this point it's a trade off of complexity for one
vs the other. Month math is easier when you store year/month/day, but
date sorting, enumerations, or createing a date from the database is
easier. We'll see...