Add events to calendar
This is a bit of a work in progress, but it does work. Before this
change, CalendarView took a callback. The callback was used, but it
seemed unnecessarily complex. To display events I think I would end up
having to do a bunch of logic in the view.

I realized the problem is once I have all the events for a particular
user over a month long period, I need to query that list of events
again. I don't want to make additional calls to the databsae, or even
one call per day. I think what I'll end up doing is making the call once
and wrapping the result in some kind of offline query object. It will
let me query specific dates in a list of events without hitting the
database.

Ironically, this may mean the return of the callback. We'll see how it
pands out.