Use EventFilterer and partials to render events
Well, the callback is back. Instead of writing more view logic in
CalendarView, this uses view partials to render events. I was able to
cut out a bunch of CalendarView logic, and it feels much more rails-y.

It does render a ton of partials (one for each day visible in the
calendar) but it doesn't seem to have a big effect on performance. We'll
see if that keeps up.