Update json gem
It was stuck at 1.8.3 because of rubocop. I've been having json
incompatibility errors with rubocop for a long time. I think what was
happening was I had rubocop installed on the system through chef
somehow. I'm not sure how it got on there, but it did. Anyway, I
switched back to system ruby, uninstalled rubocop, switched back to
2.3.1, uninstalled _all_ gems, then re-bundled and removed the version
constraint on the json gem. So far so good, but we'll see.