Add Calendar model
When adding the calendar_id to the events table, I added an index also.
I'm not sure, but I think this made an event invalid if it didn't have
an associated calendar. At least, I got an error when running tests
until I added a calendar to the event factory. ¯\_(ツ)_/¯