Add existing data and dev:prime to a calendar
For existing data, I did this as a migration. I wrapped all the events
in a transaction, then went through each of them in batches. For each
batch, I modified their calendar_id to be the calendar that I just
created. In reality this all happens as a single sql query, because
in_batches does batches of 1000 and I only have about 150 events in the
database. I did it this way for practice and good habit forming.

db/migrate/20170416021920_add_existing_events_to_calendar.rb