Add Calendar.events_for_month method
This method returns a list of events that could be displayed in a
particular month's view of the calendar. That includes events that are
outside of the month, but help round out the first or last week. E.x. if
a month starts on a Thursday, and there's an event the day before in the
previous month, we want to grab that event too.

This replaces the logic that existed in the CalendarsController.