Integrate sidebar into design
Currently this only holds sign in/up/out logic but will eventually be
used for the list of calendars.

Of note, I added a dirty hack to make the calendar horizontally
flexible. Before this hack, the calendar would be off the right edge of
the screen. This is because of minimum sizing algorithm for flex items.

I'll try to put it simply so later me will understand. Flex items have
content. That content has width. The width of that content is, by
default, the minimum size of the flex item. Apparently, this also
applies to parent items. I'm not _exactly_ sure why, but setting
min-size: 0 on the <main> element the calendar stared resizing.

I'm not sure why <main> was the clincher here. Why not the li elements
that had major amounts of text in them? Why not every flex parent beyond
that? I guess <main> kind of makes sense because it's the highest level
item that isn't <body> and <body> can't be less (or more) than 100% of
the window's width.

Anyway, that's why it's there.

More info here:
http://stackoverflow.com/questions/36247140/why-doesnt-flex-item-shrink-
past-content-size