Add basic permissions
A permission links a user to a calendar with a specific access level. At
any given time, a user has calendars they own (edit not only
events, but also add/remove other users), edit (add/remove events), or
simply view (no editing).

I chose to use a string to represent the level of access instead of a
numeric value because the order of the items shouldn't matter. The
default level is 'viewer'.