Use an array for permission levels instead of a hash
The has wasn't getting me anything execpt for being able to change the
value in the database later. So for example, I could look up all
permissions with the Permissions::LEVELS[:own] level, but why not just
look up with "owner"?

I'll have to remember the string values, but that's the same as
remembering the hash keys.