Update to git:// url for simple_form in Gemfile
This is a test to see if heroku will download a gem via https:// instead
of an ssh connection. Apparently heroku can't use the key that I
uploaded to it for connecting...

This will be solved when simple_form updates its dependencies and I can
go back to using the normal gem.