Create SignUpUser class
This is in charge of creating a user record via Monban from user params,
and also assigning a new calendar to the user with the correct
permissions.

I also had to change the default factory for a user to not create the
calendar and permisisons in the after(:create) block for testing. I
think right now certain tests will fail, but will soon be fixed when I
use this SignUpUser in the UsersController.