Use SignUpUser in UsersController and tests
Note the lack of attributes_for(:user) in the tests. This is because
FactoryGirl creates a hash of email and password_digest. However, Monban
strips the digest from the attributes. That means when it tries to
create a user, there is no password and the user isn't created.

I had to create attributes manually.