Users can add events
It's not pretty, but it works.

Now ever <div class="day"> has a link titled "New event". This goes a
page with the url: /calendars/:calendar_id/events/new?date=YYYY-MM-DD
The controller parsers the date from the date string, and creates a new
event on that calendar with the passed in title.

Of note, the controller kicks off the parsing of the date from the
params, and inserts it back into the params so it can be attached to the
event.