Add ability to edit events
This got kind of hacky. I think I will refactor so the date is more
explict in the view. I couldn't do the thing where I parse the date into
an ArDate from the parameters in the update action because there was no
date to change. So I had to move that logic directly into the create
action instead.