Create event based on day number
This is instead of trying to convert to and from a date string like
4711-01-04. I keep the date field, but only use it for user display.
The occurred_on field now holds the day_number of the date for the
event. This gets passed into the controller and doesn't need to be
converted.

I tried using a form object in a different branch and it got
complicated very quickly trying to display errors and convert the date.

This approach has one major downside of the date not being
user-changeable once it's set. Users will have to delete an event and
re-add it, which is a pretty big bummer.