Return user to the correct month
If a user creates an event in Gozran 3022, they should be returned to
Gozran 3022 after they event is created or updated. Before this commit,
it would take users back to the default month/year (Abadius 4711). Now
it uses the event that was created/updated to navigate to the month of
the calendar.