Redirect if guest user tries to view a calendar
Not signed in users can't see calendars. If a not-signed in user tries
to view a calendar, the app will redirect them to the sign in page.

One alternative I considered was to use a custom method called
authorize on the AuthenticatedController instead of Monban's
require_login. I could set the flash manually which would take away
some of the magic when Monban sets the flash. I decided not to because
it's so nice to have Monban do it's very simple logic, even if it will
be harder to find later.