Add method to check if user can own a calendar
This looks for permissions of the level 'owner' that are between a user
and a calendar.

I struggled to understand the difference between asking "what can the
user do" vs "what does a calendar allow the user to do". This
difference is important when deciding what permission levels to look
for. The former looks for permissions with only the owner level (this
commit) and the latter looks for permissions with ALL levels. Big
difference.

Ultimately, I decided on going from the perspective of the user.