Don't allow event creation if a user has been denied access
This is an unlikely scenario, but it's possible that someone revokes a
users access to edit a calendar as they're creating the event. In that
case, it should do the same thing as it does when a user tries to add
an event: send them back to the calendar page with a message that says
they can't do that.