Don't let users edit an event on a view-only calendar
At first, it don't even give them a link to edit the event on the
calendar page and only shows the title. But if they some how guess the
event id and type in the edit page url, it redirects them back to the
calendar page with a message saying they don't have access.