Refactor EventsController
Instead of dealing with an unauthorized user in the controller action,
I raise a NotAuthorized exception and handle it in the base
AuthorizedController.

I first tried to do this with a method in the events controller. It
ended up not being helpful, because the method uses a redirect_to and
so do a couple of actions. To avoid a multiple redirects error, I had
to return from the method. But I need to return from the original
controller action, not some extracted method.

I could rescue_from'd in the EventsController itself, but this seems
like it might be more general in the future? It'll be easy to move if I
need to.