Add last_edit_date to calendar
This method returns the ArDate of the last event updated on the
calendar. The idea is, you probably want to go back to the date you
were editing when navigating back to the calendar, not some arbitrary
start  date every time. So this grabs that date.

If there are no events on the calendar, it returns the default ArDate.