Take users back to the date they were editing
When a user visits a calendar, they should be taken to the date they
last changed an event on. So, if the last thing they did was change the
title of an event, it should take them to the month that that event is
in.

If no events exists yet, it will take them to the default date of
4711-01-01