Fix event spec
Since adding the automatic calendar month navigation, some event tests
broke because the test relied on the calendar starting at the default
date instead of the month last edited.

I could remove the feature I added to take users back to the same date
where they added the event. However, if I for whatever reason change
the behavior that I just created (what date does the user get taken to
on a specific calendar) I'd still want this to function the way it does
and not change.