Allow users to change the date of events
Before, I was hiding the occurred_on param as a hidden field. But I
figured out a decent way to convert the date string like `4700-02-03`
into a day_number for the params. So I always pass in the date string,
and it always gets converted to a number. This allows users to enter in
date in the event forms and thus, change the date of the event easily.