Update rubocop settings
There were a few things I didn't agree with, so I changed their
settings.

* ABC size is fine in theory, but I think it's legitimatly okay to go
above it if you feel the logic is sound.
* Similar thing with max class length. This was particularly difficult
with ArDateParser as there are a bunch of constants that take up 14+
lines. I could force those into a couple of lines, but they would harder
to read and make the code worse for the metric.
* Empty methods being on one line looks gross to me. I don't want to
think about if a method should be single or multiple lines.
* OpMethod was warning me that I when I define the >> method, the
argument should always be `other`. In my particular case, >> is
increasing a date by a number of months. I think using "months" as the
argument name is ideal here. I kept it enabled for the rest of the
project.
* The shcema is auto-generated so I don't really have control over the
format. Ignored.
* I also ignore block length on rake tasks, because if you wrap a few
tasks in a namespace, it becomes to long, but there's no other way to do
it other than maybe breaking up the file? Seemed counter productive.