Truncate long calendar names in the sidebar
For all those long, long calendar names. This is distinct from the
title length restriction on calendars because while I can see a use
case for calendars with long names and description, I can't for
displaying the whole name in the sidebar.

The other alternative is to wrap the text on multiple lines. If I do
that (and I might) I'll have to come up with a way to visually
differentiate between different calendars vs different lines of a
calendar title. Currently they are indistinguishable

fixes #61