Remove other layouts in favor of calendar layout
In making views with forms or information, everything needed about the
same amount of space. The outlier was the full calendar view since it
needs so much space.

So I created a calendar layout specifically for the Calendar's show
view and everything else can work with the Application layout. No more
need for a custom session or welcome layout.

Some styles changes slightly and now the sidebar is on every page, but
it works fine as a global navigation anyway.