Add styles to flash messages
I moved the flash to be next to the content, and in the process, moved
the content to the top of the screen. If it floats in the middle, flash
messages will push everything down to re-center. Top looks fine anyway.

I also shortened the calendar not found flash message because it was
wrapping to the next line and it looked not as good. I realize that not
every screen size will allow it to be one line, but I can do it for
many cases.

Colors inspiration from: https://dribbble.com/shots/965125-Stream-Status