Add bullet to detect N+1 queries and eagar loading
Bullet logger logs to /logs/bullet.log. I like this because I'll miss
it if it pops up in the dev log, but even if I miss it in the bullet
log, I can inspect all of the warnings later.

Footer should show up in the bottom left of the page

This was originally added so I could debug what I thought was an N+1
query when loading events and their associated calendar's permissions.
However, after trying several times at preloading, rails seems to cache
the permissions and I don't need to preload anything ¯\_(ツ)_/¯