Add session_actions to sidebar presenter
You may notice that sometimes the presenter uses `a` and sometimes it
uses `link_to`. This is because if you have two sibling tags using
link_to, only the last one is returned. If you want them both, you have
to prepend them with the text method.

My policy for now is use `link_to`, unless is has a sibling tag in
which case use `a`. We'll see how that works. They both look pretty
good.