Replace Sidebar content with presenter methods
Keynote needs to know which presenter to use. You can either pass a
class instance to it, or a symbol. If you pass a symbol, you need to
also pass an object for it to present.

In this case, I could have passed `:calendar` along with the
`@calendar` variable, which is sometimes nil. It doesn't really matter
if it's nil or not, so I could even pass `nil`.

This seemed stupid though. I passing new calendar model instead. This
will let keynote find the presenter and not worry about nil or not
instance variables. It does sort of make me feel like this could be a
helper or different presenter. _sigh_