Use new circleci 2.0 config
This involved updating my ruby version (not a bad idea anyway) and
diverting a bit from the recommended circle guides.

How I decided to configure the PGUSER:
https://hub.docker.com/r/circleci/postgres/

And I don't have a format for running rspec. This because rspec seemed
to be outputting all the little dots along with color code information
instead of the XML that it needed to output. This removes --format
option which worked on my local machine.