Add invitation model
This model is a record of every invitation a user has sent to an email
address from a calendar.

I use email address, not user, on purpose here. This is because I don't
want to expose email addresses registered in the system. If I store a
user, I'd have to associate that email with a current user, or send
back an error, which shows the inviting user which email addresses are
valid.

Having the invitation lets me keep track of all pending invitations
too. A user can see all the invitations they've ever issued. More
importantly, they can revoke invitations if they want to.