First stab at sending invites
This works by an owner of a calendar entering an email address of a
user they want to invite. This email will then show up in a list on the
same page so the owner can revoke or view all previous invites.

Eventually, the user of that email will sign in and be able to accept
the invite, looked up via their email. If no user with that email ever
exists, or the user with that email never logs in, the invite will just
sit there forever.

For a hot second there, I was also checking to make sure the email
being invited also existed in the database. However, I don't want that
because I don't want to reveal to the inviter what emails exist and
what don't. This saves the invitation regardless of what users or
emails exist in the database.