Refactor sidebar invitation link
I created a sidebar presenter. Astute readers of previous commits will
note that I attempted to create one before then refactored it out into
user and calendar presenters. What gives?

Mostly, I changed my mind. I was thinking that calling the database
(via activerecord) was a no-no. And it is, but when it's in a
presenter, it somehow feels better because it's encapsulated. I plan to
add more methods, and even transfer some methods back out of the user
and calendar presenter in the future.

For now, onward with invitations!