Ability to revoke invitations
Mostly this is just a delete action for a calendar invitation. It
starts getting a little more complicated though because invitations are
listed on the same page as the form to create one.

This is somewhat complicated because an easy way to create a model that
belongs to other models is `calendar.invitations.new`. This
automatically sets calendar_id on the invitation so you don't have to
ask the user to do it or set it in the ERB code.

The problem is, now this is in the list of invitations for that
calendar, even if it's not persisted in the database yet. So when I
list all the invitations, I get a blank invitation also.

I ended up creating the invitation more manually in the controller, but
only after also looking into creating a presenter for that page or
putting the values directly into the ERB. Neither of those options were
as clean.