Refactor partials to not assume instance variables
This seems like a better pattern to me because you have to make
everything explicit. You'll know if an object isn't available because
you had to pass all of them in.

I was working with a calendar object, so I had to rename the calendar
view helper to calendar_display.