Restrict calendar editing to calendar owners
This is done just like calendar access is. If you try to visit the edit
page for a calendar, you are thrown back to the calendar edit page with
a message saying you are not authorized to do that.

I also protected against a POST/PUT/PATCH request to change the data,
but I think that may be redundant because of CSRF protection. I need to
double check that.