Add accepted_at to invitation
This is so I can choose to display or not display certain invitations. I
considered deleting the invitations, since there's no place where I want
to display invitations that have been accepted. But then I decided to
err on the side of keeping old data around because it can be useful.
It'll be easy to delete them all later if I don't want them, but it will
be impossible to bring them back if I deleted them.

Also, time for a boolean has not updated yet so I had to grab the master
branch until a new gem version is released.