Use https link to fetch time_for_a_boolean
My heroku deploy failed when it tries to fetch the time_for_a_boolean
gem. This is because I used an ssh link but heroku could not connect to
GitHub via ssh without an ssh key.

There are a few options here.

One is to use an ssh url like git@github.com... but that requres an SSH
key, which Heroku does not have.

Another is to use a git:// url but that is considered insecure by
bundler because of a man-in-the-middle attack.

Finally, I can use an https:// link, which seems to satify both
requiremens of being able to deploy on heroku, and also being secure.