Refactor calendar edit view
I refactored this so it uses a new presenter: CalendarEditPresenter.
This has methods for the list of permissions and invitations.

Also, this tests the result of the presenters a little differently.
Instead of comparing straight html strings, this creates a
Capybara::Node::Simple and uses the normal Capybara finder matchers on
it. That's how I'm testing the output of each presenter method.