Scroll pages when they are long
This adds overflow: scroll to the <main> element of the page. There
aren't a ton of pages that can be particularly long right now, but there
are some. the edit calendar page is one because it has permissions,
invitations, and an invitation form, along with the actual calendar
name.

The page was already scrolling, but it would also scroll the sidebar up
when scrolling the main content down. This keeps the sidebar where it
is, and makes everything inside <main> scroll instead.