Re-style month navigation
This brings the next and previous links to center to flank the
Month/Year and re-styles them to be graphical buttons with < and >
arrows.

I had to change the viewport slighly on the previous arrow because it
just looked a little off center. Not sure why the "next" arrow looked
centered to me.