Fix ArDate attribute
When I built this, I don't think I had a good understanding of exactly
what it was doing. I don't think my understanding it perfect yet, but
it's better now.

When a value that is mapped to an ArDate is passed to an ActiveRecord it
calls serialize to make sure it's turned into an integer (the day
number) that can be stored in the database. When it is coming back from
the database, it maps this integer back to an ArDate by calling either
cast or deserialize. I think.

I'm not sure why cast is needed along with deserialize.

I do know that I was casting the ArDate string wrong. I was just turning
it into an integer which was trucating a date like "4711-07-23" to just
4711, which is a very low day number compared to the 1719739 it should
be. Now it detects a String and uses ArDateParser.from_date_string
instead of to_i.

I also cleaned up a weird if statement.