Add the ability to set the current date on a calendar
This allows users to click "Set Today" which is in the place where "New
Event" used to be. Set today will send a patch requset to the Calendars
controller and hit the current_date action with the current date. It
only permits current_date as a parameter.

I also considered just passing the single parameter to update, but I
liked the specificity of an action that _only_ handled setting the
current day.