Rename week_days to week_dates
I've been struggling for a bit with the difference between date and
day. The rule of thumb I've come up with is: if it involves a month or
day in _any_ way, it's a date. Otherwise, it's a day.

Example: this method `week_dates` represents a sequence of seven dates
within a month/year. However, `day_names_header` is Moonday - Sunday no
matter what month or year it is.