Return empty <ul> when there are no calendar events
The `events` method on the CalendarPresenter was returning an empty
array string "[]" when there were no events on that day. This was not a
good look for the calendar as it would have a "[]" on each day that
didn't have an event yet. Now it returns just the empty `ul`