Scroll events when there are too many
This scrolls the ul.events element when there are too many to
vertically fit in the calendar date div. I also considered scrolling
the entire contents of the date div, but in that case it was not
obvious that you could scroll and it looked weird.

This implimentation has the downside of the events looking pretty
squished on everything but a big screen. Although, maybe that's okay
because if you're interacting with that day you'd should see the
actions over the content.