Make system tests work again
Turns out when I added the capybara host bit, it broke system tests. I
think this is because when the test starts up, it tries to connect to
whatever host capybara has set, and then just can't connect and
timeout's.

Note that one tests still fails: User decides not to add an event and
is redirected back to their calendar in
spec/system/user_adds_an_event_spec.rb This is because current_url and
the URL path helpers still don't match up. I'll figure out how to fix
this soon, but I wanted this fix to be a singular change.