Fix failing test
This test was checking to make sure the user is directed back to the
date in the calendar that they were previously editing, as opposed to
some default date.

I was using current_url and checking that against a url helper. But in
system tests, the host doesn't match. I don't really care about the
host in this case, so I switch to using a path helper and the
have_current_path helper which I just learned about.