Set the Event hidden_at default to now() instead of Time.current
The problem was Time.current was evaluated when the migration ran so it
was always some particular date, as opposed to the date the event was
created.

This will use postgresql's `now()` function to set the date to whenever
the INSERT transaction runs. `CURRENT_TIMESTAMP` is also an option, but
`now()` is what gets inserted into the schema.rb file and it's
shorter/nicer looking IMO.

See the end of this section for more discussion:
https://www.postgresql.org/docs/9.5/static/functions-datetime.html#FUNCT
IONS-DATETIME-CURRENT