Remove now default from Event hidden_at
When creating a new event (through ActiveRecord or FactoryBot) I would
sometimes want to set the hidden_at attribute to nil so the event was
not hidden or, visible. The problem (I think) is that postgresql was
seeing this attribute, as nil, and then setting it's default to the
current time. So the event would be created, then a default for
hidden_at would be set, making the event visible.

I removed the default and I'll make it so the view enables the hidden
attribute by default.