Hidden checkbox will reflect the hidden state of the event
I had copied and pasted the checkbox from the new form. In the new
form, I want it to default to checked, but in the edit form, I want it
to reflect whatever the event is set to.

I wanted to use the have_unchecked_field matcher in the test, but for
whatever reason, it could not find the element. So I manually found it
and used the be matcher on `checked?`