Associate permissions with associations
This was a difficult decision

The impetus here was that there is a uniqueness validation on
invitations that doesn't let someone invite the same email twice to a
calendar. Since invitation allow the creation of permissions, if there
are multiple duplicate invitations, there can be multiple duplicate
permissions.

However, this makes it impossible to invite someone after you've revoked
their permission. Let's say you kicked them out of your game but then
invited them back. You can't do this because the invitation sticks
around and the uniqueness validation prevents it from being created,
even if the old permission is deleted.

That's why I connected the invitation to the permission, so I can delete
them together.