Remove invitations when revoking permissions
This ended up being pretty simple because of some rails helpers.

First was linking new permissions to invitations. When a model
`has_one` other model (for example `permission`), you can use
`build_association` method (for example `build_permission`) to
automatically set the `invitation_id`.

Second was making sure the invitation was deleted along with the
permission. This was as simple as adding the `dependent: :destroy`
option. Now when a Permission is destroyed, it destoy's it's associated
invitation.