Wrap initial user/calendar creation in a transaction
If this isn't wrapped in a transaction, then there's not much to stop
the calendar and associated permission from being saved (except for the
new user_id validation on permission). But this way, if the user isn't
valid, the transaction will fail, and no calendar, permission, or user
will be created. The same is true if any other method runs in to
trouble being created.