Pass the SignUpUser params in the new method instead of perform
The whole pattern of using a class method that calls new and then an
instance method by the same name is great. But it makes more sense to
pass in the user params to the class and store it there.

Calling the class method works exactly the same after this change.