Take users to "today" on a calendar by default
Before I had the concept of "today" on a calendar via the
`current_date` attribute, I used the date of the last edited event as a
stand in. However, now that I have `current_date`, I can use that
instead.

This still take users back to the month/year that they were doing event
creation/modifying on. The create and update controller actions still
pass the date param. But if there is no param, the calendar uses its
current_date.

I had to update a test that assumed navigating to the calendar would
start display the month where the calendar was created. This used to be
true, but now it's not.